feat(animate)!: add `max_output_steps` option to cursor path generators
The option is added in default, `gen_path.line`, and `gen_path.angle`
generators. Default value is 1000.

This is breaking behavior (as previously there were no constraints)
mostly technically because on big cursor jumps it was already invisible
in parts. This is because timer in `animate()` can work only in 1 ms
increments with steps sooner than 1 ms after latest drawn are executed
immediately. This was a performance issue as it meant that visualization
extmarks were added and removed without any visible gain.

Resolve #1374